+++ src/album_list/mod.rs
@@ -2,30 +2,23 @@ use ::ascii_tree::write_tree;
 
 use crate::album_list::ascii_tree::build_tree;
 use crate::db::repo::album::AlbumRepository;
-use crate::library::PhotosLibrary;
 use crate::model::album::Album;
 use crate::model::FromDbModel;
+use crate::result::PhotosExportResult;
 
 mod ascii_tree;
 
-pub fn print_album_tree(library_path: String) {
-    let library = PhotosLibrary::new(library_path);
-    let album_repository = AlbumRepository::new(library.db_path());
+pub fn print_album_tree(db_path: String) -> PhotosExportResult<()> {
+    let album_repository = AlbumRepository::new(db_path);
 
-    let db_albums = album_repository.get_all().unwrap_or(Vec::new());
+    let db_albums = album_repository.get_all()?;
 
-    let album_result: Result<Vec<_>, _> = db_albums
+    let albums: Vec<Album> = db_albums
         .iter()
         .map(|a| {
             Album::from_db_model(&a)
         })
-        .collect();
-
-    if album_result.is_err() {
-        eprintln!("At least one album could not be converted from DB model");
-        return;
-    }
-    let albums = album_result.unwrap();
+        .collect::<Result<Vec<Album>, String>>()?;
 
     let tree = build_tree(&albums);
 
@@ -33,4 +26,6 @@ pub fn print_album_tree(library_path: String) {
     let _ = write_tree(&mut ascii_tree, &tree);
 
     println!("{}", ascii_tree);
+
+    Ok(())
 }