+++ CHANGELOG.md
@@ -4,7 +4,10 @@ The following is a list of changes made to the project in reverse chronological 
 
 ## `0.3.0-snapshot`
 
-- Rename database models to be better distinguishable from the library models.
+- Improve error handling and error messages
+  - Errors are now exported to an error log file in the export directory
+  - Internal error handling has been improved
+- Rename database models to be better distinguishable from the library models
 
 ## `0.2.0`